Re-organize folders and paths
