+++ data/albums.json
@@ -1,13 +1,13 @@
 [
   {
-    "publish_date": "2022/02/16",
+    "publish_date": "2019/08/08",
     "on_spotify": "true",
     "genre": {
       "name": "rock"
     },
     "author": {
       "first_name": "rock",
-      "last_name": "rco"
+      "last_name": "rock"
     }
   }
 ]
