+++ lib/create_book.rb
