+++ lib/label.rb
