+++ lib/music_album.rb
@@ -1,5 +1,6 @@
-require_relative './item'
-require_relative './genre'
+require_relative 'item'
+# require_relative '../item'
+require_relative 'genre'
 require 'date'
 
 # Album class
